refactor LogToConsolePrinter class
