+++ jaeger-spark-dependencies-elasticsearch/src/test/java/io/jaegertracing/spark/dependencies/elastic/ElasticsearchDependenciesDockerJobTest.java
@@ -4,6 +4,7 @@
  */
 package io.jaegertracing.spark.dependencies.elastic;
 
+import io.jaegertracing.spark.dependencies.LogToConsolePrinter;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.utility.DockerImageName;
